renamed a few more things
